Add watch for component files, refactor.
