+++ lib/Bootstrapper.js
@@ -32,7 +32,8 @@
 
 var util = require('util'),
 	path = require('path'),
-	BrowserBundleBuilder = require('./BrowserBundleBuilder'),
+	BrowserBundleBuilder = require('./builders/BrowserBundleBuilder'),
+	BootstrapperBuilder = require('./builders/BootstrapperBuilder'),
 	FileWatcher = require('./FileWatcher'),
 	StoreFinder = require('./finders/StoreFinder'),
 	ComponentFinder = require('./finders/ComponentFinder'),
@@ -79,6 +80,9 @@ Bootstrapper.prototype.configure = function (configObject, locator) {
 
 	locator.register(
 		'browserBundleBuilder', BrowserBundleBuilder, configObject, true
+	);
+	locator.register(
+		'bootstrapperBuilder', BootstrapperBuilder, configObject, true
 	);
 	locator.register('fileWatcher', FileWatcher, configObject, true);
 	locator.register('storeFinder', StoreFinder, configObject, true);
