+++ lib/FileWatcher.js
@@ -32,11 +32,12 @@
 
 module.exports = FileWatcher;
 
-var Gaze = require('gaze').Gaze;
+var chokidar = require('chokidar');
 
-var GLOB_OPTIONS = {
-	interval: 1000,
-	debounceDelay: 5000
+var OPTIONS = {
+	//ignoreInitial: false,
+	cwd: process.cwd(),
+	ignorePermissionErrors: true
 };
 
 /**
@@ -44,15 +45,15 @@ var GLOB_OPTIONS = {
  * @constructor
  */
 function FileWatcher() {
-	this._watchObjects = [];
+
 }
 
 /**
- * Current list of watched objects.
- * @type {Array<Gaze>}
+ * Current watch object.
+ * @type {Chokidar}
  * @private
  */
-FileWatcher.prototype._watchObjects = null;
+FileWatcher.prototype._watcher = null;
 
 /**
  * Watches after specified resources and invokes handler on changes.
@@ -60,36 +61,26 @@ FileWatcher.prototype._watchObjects = null;
  * @param {Function} handler Handler of all changes.
  */
 FileWatcher.prototype.watch = function (globs, handler) {
-	var self = this,
-		lastTriggered = 0,
-		safeHandler = function () {
-			var now = Date.now();
-			if (lastTriggered !== 0 &&
-				now - lastTriggered <= GLOB_OPTIONS.debounceDelay) {
-				return;
-			}
-			lastTriggered = now;
-			handler();
-		};
+	if (this._watcher) {
+		return;
+	}
 
-	var watcher = new Gaze(globs, GLOB_OPTIONS,  function () {
-		this.on('error', function (error) {
+	var self = this;
+	this._watcher = chokidar.watch(globs, OPTIONS);
+	this._watcher
+		.on('change', handler)
+		.on('error', function (error) {
 			self._eventBus.emit('error', error);
-		})
-			.on('added', safeHandler)
-			.on('changed', safeHandler)
-			.on('deleted', safeHandler)
-			.on('renamed', safeHandler);
-	});
-	this._watchObjects.push(watcher);
+		});
 };
 
 /**
  * Removes all watches from the list.
  */
 FileWatcher.prototype.unwatch = function () {
-	this._watchObjects.forEach(function (watch) {
-		watch.close();
-	});
-	this._watchObjects = [];
+	if (!this._watcher) {
+		return;
+	}
+	this._watcher.close();
+	this._watcher = null;
 };
